Изменен принцип организации интерфейса взаимодействия с внешней моделью. Функционал перенесён в класс Optimizer (つ╥﹏╥)つ
