--- Interfaces/InternalTestInterface/__init__.py
@@ -1,2 +1,0 @@
-from .InternalTestInterface import InternalTestInterface
-
