--- Interfaces/__init__.py
@@ -1,4 +1,0 @@
-from Interfaces.BaseInterface import OptimisationTypes
-
-from Interfaces.InternalTestInterface import InternalTestInterface
-
